requirements2.txt
bidict==0.23.1
cffi==1.17.1
click==8.1.7
colorama==0.4.6
cryptography==41.0.3
Flask==2.2.5
Flask-Login==0.6.2
Flask-SocketIO==5.3.4
freeze-requirements==0.5.3
h11==0.14.0
itsdangerous==2.2.0
Jinja2==3.1.2
MarkupSafe==3.0.2
pycparser==2.22
pycryptodome==3.21.0
python-engineio==4.10.1
python-socketio==5.11.4
sh==2.1.0
simple-websocket==1.1.0
six==1.16.0
Werkzeug==2.3.7
wsproto==1.2.0
